% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformTrain.R
\name{consolidateModels}
\alias{consolidateModels}
\title{Copy results from the model file to the main object.}
\usage{
consolidateModels(isot, id = 1:length(isot@domain))
}
\arguments{
\item{isot}{An \code{\link{IsoformTrain}} object.}
}
\description{
Copy results from the model file to the main object.
}

