% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-EpiDatatype.R
\name{convert2bin}
\alias{convert2bin}
\title{Convert alignment files to bin counts (both treatment and control), supports tagAlign and bam for now.}
\usage{
convert2bin(epidt)
}
\arguments{
\item{epidt}{EpiDatatype (need link) object}
}
\description{
Convert alignment files to bin counts (both treatment and control), supports tagAlign and bam for now.
}

