% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformTrain.R
\name{differentialAnalysis}
\alias{differentialAnalysis}
\title{Run differential analysis for two conditions (in parallel)}
\usage{
differentialAnalysis(isot, data.type, bam.cond1, bam.cond2, n.perm,
  mix.prop = 0, chrlen.file, bin.width, n.cores)
}
\arguments{
\item{isot}{IsoformTrain object}
}
\description{
Run differential analysis for two conditions (in parallel)
}

