% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eia_cat}
\alias{eia_cat}
\title{This function download EIA categories.}
\usage{
eia_cat(api_key = "34117E2A8A50B13CAC6A3C5C8BDC89CF", category_id = NULL)
}
\arguments{
\item{api_key}{Character value. Your API key.}

\item{category_id}{Numerical value. Category.}
}
\description{
This function download EIA categories.
}

