% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{eia_clear_cache}
\alias{eia_clear_cache}
\alias{eia_clear_dir}
\alias{eia_clear_metadata}
\alias{eia_clear_data}
\alias{eia_clear_facet}
\title{Clear API results cache}
\usage{
eia_clear_cache()

eia_clear_dir()

eia_clear_metadata()

eia_clear_data()

eia_clear_facet()
}
\description{
Reset the results of API calls that are currently cached in memory.
}
\details{
\code{eia_clear_cache} clears the entire cache. The other functions clear the cache associated with specific endpoints.
}
\examples{
\dontrun{
key <- Sys.getenv("EIA_KEY") # your stored API key
system.time(eia_dir(key))
system.time(eia_dir(key))
eia_clear_cache()
system.time(eia_dir(key))
}
}
