% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eia_date_parse}
\alias{eia_date_parse}
\title{A function to parse dates from series}
\usage{
eia_date_parse(series_list, format_character)
}
\arguments{
\item{series_list}{A list of requested series.}

\item{format_character}{A vector with a character defining the date formats}
}
\description{
A function to parse dates from series
}

