% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eia_date_yq}
\alias{eia_date_yq}
\title{A function to parse quarterly dates from EIA}
\usage{
eia_date_yq(qstr, first = TRUE)
}
\arguments{
\item{qstr}{Character vector with the EIA format yyyyQq}

\item{fist}{Logical. Returns either the first or the last date in the quarter.}
}
\description{
A function to parse quarterly dates from EIA
}

