% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{eia_directories}
\alias{eia_directories}
\alias{eia_subdirectories}
\alias{eia_dirs}
\alias{eia_parent_cats}
\alias{eia_child_cats}
\title{EIA categories}
\usage{
eia_directories(dir = NULL, tidy = TRUE, cache = TRUE, key = eia_get_key())

eia_subdirectories(dir, cache = TRUE, key = eia_get_key())

eia_dirs(id = NULL, tidy = TRUE, cache = TRUE, key = eia_get_key())

eia_parent_cats(id, tidy = TRUE, cache = TRUE, key = eia_get_key())

eia_child_cats(id, tidy = TRUE, cache = TRUE, key = eia_get_key())
}
\arguments{
\item{dir}{character, directory path, if \code{NULL} then the API root directory.}

\item{tidy}{logical, return a tidier result. See details.}

\item{cache}{logical, cache result for duration of R session using memoization. See details.}

\item{key}{API key: character if set explicitly; not needed if key is set globally. See \code{eia_set_key()}.}
}
\value{
for \code{eia_directories}, a tibble data frame (or a less processed list, or character, depending on \code{tidy} value); others functions return a tibble data frame.
}
\description{
Obtain EIA directories.
}
\details{
By default, additional processing is done to return a list containing tibble data frames.
Set \code{tidy = FALSE} to return only the initial list result of \code{jsonlite::fromJSON}.
Set \code{tidy = NA} to return the original JSON as a character string.

Set to \code{cache = FALSE} to force a new API call for updated data.
Using \code{FALSE} always makes a new API call and returns the result from the server.
\code{TRUE} uses memoization on a per R session basis, caching the result of the function call in memory for the duration of the R session.
You can reset the entire cache by calling \code{eia_clear_cache}.

\code{eia_subdirectories} returns only the immediate subcategories under a given parent.
This is a wrapper around \code{eia_directories} and always return a tibble data frame.
}
\examples{
\dontrun{
# use eia_set_key() to store API key
eia_directories()
eia_subdirectories("electricity")
eia_subdirectories("electricity/rto")
}
}
