% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{eia_updates}
\alias{eia_updates}
\title{EIA data updates}
\usage{
eia_updates(...)
}
\value{
character sting, defunct message
}
\description{
NOW DEFUNCT - FEATURE NO LONGER PROVIDED WITH APIv2.
}
\examples{
eia_updates()
}
