% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformMod.R
\name{findIsoforms}
\alias{findIsoforms}
\title{Main function to run isoform searching algorithm}
\usage{
findIsoforms(isom, k, pilot.rep, pilot.iter, pilot.tol, max.iter, tol,
  n.cores = 1)
}
\arguments{
\item{isom}{\code{\link{IsoformMod-class}} object}

\item{k}{Number of isoforms to start with}

\item{pilot.rep}{Number of random starting points to begin with}
}
\description{
Main function to run isoform searching algorithm
}
