% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-EpiDatatype.R
\name{generateMatrices}
\alias{generateMatrices}
\title{Generate count matrices for each chromosome, each matrix is saved as a separate .rds files in the data directory.
Requires that bin counts be summarized from the alignment files by running \code{\link{convert2bin}}.}
\usage{
generateMatrices(epidt)
}
\arguments{
\item{epidt}{A \code{\link{EpiDatatype}} object.}
}
\description{
Generate count matrices for each chromosome, each matrix is saved as a separate .rds files in the data directory.
Requires that bin counts be summarized from the alignment files by running \code{\link{convert2bin}}.
}
