% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-EpiDatatype.R
\name{generateMatrix}
\alias{generateMatrix}
\title{Generate counts matrix per chromosome. Results are saved as .rds files in data directory.}
\usage{
generateMatrix(epidt)
}
\arguments{
\item{epidt}{EpiDatatype Object}
}
\description{
Generate counts matrix per chromosome. Results are saved as .rds files in data directory.
}

