% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-EpiDatatype.R
\name{getSizeFactors}
\alias{getSizeFactors}
\title{Only used when file type given is binary bin counts already, otherwise, use \code{\link{convert2bin}} will calculate the size factors.}
\usage{
getSizeFactors(epidt)
}
\arguments{
\item{epidt}{A \code{\link{EpiDatatype}} object.}
}
\description{
Only used when file type given is binary bin counts already, otherwise, use \code{\link{convert2bin}} will calculate the size factors.
}
