% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{image.na}
\alias{image.na}
\title{wrapper of image (link) that allows plotting NAs}
\usage{
\method{image}{na}(z, zlim, col = brewer.pal(9, "Blues"),
  na.color = grey.colors(1, 0.95), outside.below.color = "black",
  outside.above.color = "white", rowsep = NULL, colsep = NULL,
  sepcolor = grey.colors(1, 0.95), sepwidth = 0.02, cellnote = F,
  digit.format = "\%0.2f", text.col = "black", ...)
}
\description{
wrapper of image (link) that allows plotting NAs
}

