% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IsoformTrain-class}
\alias{IsoformTrain-class}
\title{The main class for training an isoform model}
\description{
The main class for training an isoform model
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the analysis.}

\item{\code{data.types}}{Name of the epigenomic data types.}

\item{\code{epidt}}{A list of \code{\link{EpiDatatype}} objects.}

\item{\code{domain}}{A \code{\link{GenomicRanges}} object specifying the local domains over which to individual isoform models are to be applied.}

\item{\code{domain.list}}{A list of \code{\link{DomainMod}} objects, corresponding to each domains.}

\item{\code{mod.dir}}{The directory for domain model files.}

\item{\code{cell.types}}{Names of the cell types used for the analysis.}

\item{\code{bin.width}}{The length of bins to be used for summarizing counts.}

\item{\code{chrlen.file}}{A text file with the length of chromosomes of the target genome.}

\item{\code{chr.count}}{The number of chromosomes there is.}
}}

