% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotIsoformModel}
\alias{plotIsoformModel}
\title{Plot the isoform fitting results, called by \code{\link{plotDomain}}}
\usage{
plotIsoformModel(isom, zlim, col.sel = NULL, main = "", cell.names = "")
}
\arguments{
\item{isom}{IsoformMod object}

\item{zlim}{Range of values in the obeserved matrix}

\item{col.sel}{Indices of selected columns to plot, default to all columns.}

\item{main}{Title of plot}

\item{cell.names}{Row names}
}
\description{
Plot the isoform fitting results, called by \code{\link{plotDomain}}
}

