% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformTrain.R
\name{runDifferentialAnalysis}
\alias{runDifferentialAnalysis}
\title{Run differential analysis for two treatment conditions (in parallel)}
\usage{
runDifferentialAnalysis(isot, data.type, bam.cond1, bam.cond2, n.perm = 2000,
  chrlen.file, bin.width, n.cores, id = 1:length(isot@domain))
}
\arguments{
\item{isot}{IsoformTrain object}
}
\description{
Run differential analysis for two treatment conditions (in parallel)
}

