% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformTrain.R
\name{runModelFit}
\alias{runModelFit}
\title{Run model fit for a selected set of domains}
\usage{
runModelFit(isot, domain.id, ...)
}
\description{
Run model fit for a selected set of domains
}
\section{Slots}{

\describe{
\item{\code{isot}}{An \code{\link{IsoformTrain}} object.}

\item{\code{domain.id}}{A vector of the indices of domains to run model fit.}
}}

