% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformTrain.R
\name{setupDomainModel}
\alias{setupDomainModel}
\title{Setup the domain model}
\usage{
setupDomainModel(isot, chrlen.file, bin.width, mod.dir,
  data.types = isot@data.types)
}
\description{
Setup the domain model
}

