% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-EpiDatatype.R
\name{convert2bin}
\alias{convert2bin}
\title{Convert alignment files to bin counts (both treatment and control), outputs are binary .bin files, currently supports tagAlign and bam files as inputs.
The normalizing size factors will also be calculated during the conversion.}
\usage{
convert2bin(epidt)
}
\arguments{
\item{epidt}{A \code{\link{EpiDatatype}} object.}
}
\description{
Convert alignment files to bin counts (both treatment and control), outputs are binary .bin files, currently supports tagAlign and bam files as inputs.
The normalizing size factors will also be calculated during the conversion.
}

