% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\docType{package}
\name{eia}
\alias{eia}
\alias{eia-package}
\title{eia: EIA API wrapper}
\description{
This package provides API access to data from the US
\href{https://www.eia.gov/}{Energy Information Administration} (EIA).
}
