% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{eia_cats}
\alias{eia_cats}
\alias{eia_child_cats}
\alias{eia_parent_cats}
\title{EIA categories}
\usage{
eia_cats(api_key, id = NULL, tidy = TRUE)

eia_child_cats(api_key, id)

eia_parent_cats(api_key, id)
}
\arguments{
\item{api_key}{character, API key.}

\item{id}{integer, category ID.}

\item{tidy}{logical, return a tidier result. See details.}
}
\value{
a list for \code{eia_cats}; others functions return a tibble data frame.
}
\description{
Obtain EIA categories.
}
\details{
Set \code{tidy = FALSE} to return only the initial result of \code{jsonlite::fromJSON}.
By default, additional processing is done to condense some list items into data frames and all data frames are converted to tibble data frames.

\code{eia_child_cats} returns only the immediate child categories. \code{eia_parent_cats} returns all parents.
}
\examples{
\dontrun{
key <- readRDS("key.rds") # your stored API key
eia_cats(key)

eia_child_cats(key, 389) # immedate children
eia_parent_cats(key, 742) # all parents
}
}
