% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{eia_clear_cache}
\alias{eia_clear_cache}
\alias{eia_clear_cats}
\alias{eia_clear_series}
\alias{eia_clear_geoset}
\title{Clear API results cache}
\usage{
eia_clear_cache()

eia_clear_cats()

eia_clear_series()

eia_clear_geoset()
}
\description{
Reset the results of API calls that are currently cached in memory.
}
\details{
\code{eia_clear_cache} clears the entire cache. The other functions clear the cache associated with specific endpoints.
}
\examples{
\dontrun{
key <- Sys.getenv("EIA_KEY") # your stored API key
system.time(eia_cats(key))
system.time(eia_cats(key))
eia_clear_cache()
system.time(eia_cats(key))
}
}
