% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoset.R
\name{eia_geoset}
\alias{eia_geoset}
\title{EIA geoset data}
\usage{
eia_geoset(key, id, region, start = NULL, end = NULL, n = NULL,
  tidy = TRUE, cache = TRUE)
}
\arguments{
\item{key}{character, API key.}

\item{id}{character, geoset series ID, may be a vector. See details.}

\item{region}{character, region ID, may be a vector. Data available for the intersection of \code{id} and \code{region} is returned.}

\item{start}{start date. Providing only a start date will return up to the maximum 100 results if available.}

\item{end}{end date. Providing only an end date will a single result for that date.}

\item{n}{integer, length of series to return ending at most recent value or at \code{end} date if also provided. Ignored if \code{start} is not \code{NULL}.}

\item{tidy}{logical, return a tidier result. See details.}

\item{cache}{logical, cache result for duration of R session using memoization. See details.}
}
\value{
a tibble data frame
}
\description{
Obtain EIA geoset data.
}
\details{
\code{id} may be a vector. This should only be done with \code{tidy = TRUE} if the tidied results can be properly row bound.
The geoset API calls allow multiple regions, but expect a single series ID. This function will one API call per series ID.
There is an expectation of similarly formatted series that can be row bound. If the IDs are for differently structured data that cannot be tidily row bound,
you may as well make separate requests since each requires a unique API call either way.

Set \code{tidy = FALSE} to return only the initial result of \code{jsonlite::fromJSON}.
By default, additional processing is done to return a tibble data frame.

Set to \code{cache = FALSE} to force a new API call for updated data.
Using \code{FALSE} always makes a new API call and returns the result from the server.
\code{TRUE} uses memoization on a per R session basis, caching the result of the function call in memory for the duration of the R session.
You can reset the entire cache by calling \code{eia_clear_cache()}.
}
\examples{
\dontrun{
key <- Sys.getenv("EIA_KEY") # your stored API key
id <- paste0("ELEC.GEN.ALL-99.", c("A", "Q", "M"))
region <- c("USA-CA", "USA-NY")

eia_geoset(key, id[1], region[1], start = 2016)
eia_geoset(key, id[2], region, n = 10)
eia_geoset(key, id[3], region[2], end = 2016, n = 10)

# multiple series counted as a single API call
x <- eia_geoset(key, id, region[1], end = 2016, n = 10)
x$data[[1]]
}
}
\seealso{
\code{\link{eia_clear_cache}}
}
