% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{eia_series_metadata}
\alias{eia_series_metadata}
\alias{eia_series_updated}
\alias{eia_series_dates}
\alias{eia_series_range}
\title{EIA series metadata}
\usage{
eia_series_metadata(key, id, cache = TRUE)

eia_series_updated(key, id, cache = TRUE)

eia_series_dates(key, id, cache = TRUE)

eia_series_range(key, id, cache = TRUE)
}
\arguments{
\item{key}{character, API key.}

\item{id}{character, series ID, may be a vector.}

\item{cache}{logical, cache result for duration of R session using memoization.}
}
\value{
a tibble data frame
}
\description{
Make a small request to obtain a data frame containing metadata.
}
\details{
Dates are provided in \code{eia_series_dates} for the convenience of working with the EIA date string format;
for example: maintaining order, generating sequences, computing intervals,
and other operations that work well with dates but would be difficult using arbitrary strings.
Keep in mind that of course these are not real dates, in the sense that you cannot map a year to a specific date.
}
\examples{
\dontrun{
key <- Sys.getenv("EIA_KEY") # your stored API key
id <- paste0("ELEC.CONS_TOT_BTU.COW-AK-1.", c("A", "Q", "M"))

eia_series_metadata(key, id)
eia_series_updated(key, id)
eia_series_dates(key, id)
eia_series_range(key, id)
}
}
