% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{eiadate}
\alias{eiadate}
\alias{eiadate_to_date}
\alias{date_to_eiadate}
\alias{eiadate_to_date_seq}
\title{EIA date parsing}
\usage{
eiadate_to_date(x)

date_to_eiadate(x, date_format = c("A", "Q", "M"))

eiadate_to_date_seq(start, end)
}
\arguments{
\item{x}{character, EIA date string; character or date object for for regular dates. Dates should be in year-month-day order.}

\item{date_format}{EIA date format: "A", "Q", "M". These stand for annual, quarterly, monthly.}

\item{start}{start EIA date or date.}

\item{end}{end EIA date or date.}
}
\description{
Helper functions for manipulating and converting between regular year-month-day date strings and EIA date string notation.
}
\details{
There is no reason to mix EIA date formats in this context. Functions that take EIA date strings expect a consistent format.
}
\examples{
eiadate_to_date(c("201803", "201804"))

date_to_eiadate("2018-05-14", "A")
date_to_eiadate("2018-05-14", "Q")
date_to_eiadate("2018-05-14", "M")

(x <- eiadate_to_date_seq("2018Q1", "2018Q4"))
date_to_eiadate(x)
}
