% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-IsoformMod.R
\name{findIsoforms}
\alias{findIsoforms}
\title{Main function to run isoform searching algorithm}
\usage{
findIsoforms(isom, k = 15, pilot.rep = 300, pilot.max.iter = 30,
  pilot.tol = 0.001)
}
\arguments{
\item{isom}{IsoformMod object}

\item{k}{Number of isoforms to start with}

\item{pilot.rep}{Number of random starting points to begin with}
}
\description{
Main function to run isoform searching algorithm
}

