% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotComponents}
\alias{plotComponents}
\title{Plot the distribution of number of isoforms across domains}
\usage{
plotComponents(isot, domain.id = 1:length(isot@domain))
}
\description{
Plot the distribution of number of isoforms across domains
}

